refactor(style-dictionary): extract semantic contract generator module
